Fix overactivePollingCausedByIdle when pipelining is requested.
BOSHClient will request pipelining (hold > 1) by default.  Make
sure this test disables it, by setting hold='1' in the session
creation response.  This happens normally when using
getSessionCreationResponse().

Reduce the value of 'polling', so this test completes more quickly.

Respond with requests='2', not '1'. It doesn't make sense for
a server to say <body requests='1' hold='1'>; that's not a usable
configuration, since you'll never be able to send a request until
you get a response to something else.  This should be prohibited
by spec.  requests='2' still exercises this test; with a value of
2, the empty request can't be sent until 'polling' seconds elapses.