Implement receive windowing.
By using a hold value larger than 1, we can increase the number of packets
the server can send to us before it has to wait for a new request.  XEP-0124
mentions this in section 4.

This means:

- When we send empty requests, we send enough to have as many exchanges held
as 'hold' specifies.  If the server specifies hold='1', this means we'll still
only send one empty request.
- The requested value of 'hold' is currently 3.  This is a tunable.
- Once we schedule a request, keep it scheduled instead of clearing it on every
received packet, unless we explicitly need to reschedule it when a response to
a pause request is received.
- Refactor empty requests a little, to ensure that if a pause request is sent
while sendEmptyRequests is running in a thread, it never sends an empty request
after we've asked for a pause.  That would wake the connection back up.

Note that an empty try ... finally has been left in send() to avoid reindenting
a large block of code, to avoid merge headaches.