Partially merge code paths for ACK resends and RBC resends.
processResponseAcknowledgementReport returns the same type of
list that we construct for isRecoverableBindingCondition, so
they can follow the same code path.

This also may fix a bug that I havn't spent the time to check:
processResponseAcknowledgementReport added the resulting exchange
to exchanges, which was also done by the caller, which probably
resulted in it being added twice.