Handle unexpected terminations.
When request acks are not supported by the server, track sent requests
in pendingRequestAcks anyway.  This tracks which responses we need to
resend after a reconnection.

Both fatal and recoverable disconnections use fireConnectionClosedOnError;
to detect whether a connection can be reconnected, call
isRecoverableConnectionLoss().

The connection to the server will not be reattempted until the user
explicitly calls reestablishConnection().  Attempts to send requests
while recoverably disconnected will block.

For low-level specifics, see TestReconnection.