SocketFactories are hard to work with; provide a simpler interface.
It's hard to layer SocketFactories to provide both TLS and proxy
support.  Simplify this: instead of taking two SocketFactories,
one for plain and one for TLS, take a single SocketFactory that we
always use to establish a connection, and a simple SSLConnector
interface to convert the socket to an SSLSocket if it's an HTTPS
connection.