Simplify BOSHClientConfig.
It takes too much code to implement a single preference, because
there's a lot of duplication.

Rework this by using a BOSHClientConfig object to store the configuration
a Builder is creating.  Builder.build() returns a clone of that config;
Builder.create clones a config as a template.  Only BOSHClientConfig
objects which are held by a Builder can be modified; the one returned
by build() is immutable, because no Builder holds it.

This way, each setting only needs to be declared once, given a getter,
setter and included in the clone ctor.