Fix NPE if disconnect() is called after the session is already disposed.
close() is called asynchronously to stop the connection; other threads
may still be trying to use it.  BOSHException should be thrown to report
this cleanly.  This should happen for disconnect() calls as well; if a
caller is attempting to close the connection cleanly and it's already
closed, it should be reported as an error.