Allow calling attemptReconnection at any time.
Rather than only attempting to reconnect when we know we're
disconnected, allow calling attemptReconnection at any time to close
and reestablish all HTTP connections.  This allows the caller to
reassert the connection based on external data, such as system
connectivity notifications (eg. Android's ConnectivityManager), rather
than having to wait for connections to fail.  This can also be used
to restart a reconnection attempt.